requirements.txt
streamlit
pandas
matplotlib
plotly.express
numpy
seaborn